Refactor DDQN and agent classes to replace RunnerState and EvalRunnerState with new Runner and EvalRunner classes
