+++ jax_imprl/__init__.py
@@ -1,7 +1,2 @@
-from .agents.primitives.agent import (
-    Agent,
-    TransitionTuple,
-    RunnerState,
-    EvalRunnerState,
-)
+from .agents.primitives.agent import Agent, TransitionTuple
 from .agents.primitives.MLP import MLP
